requirements.txt
# Ядро фреймворку
Flask==3.0.0
Flask-SQLAlchemy==3.1.1

# Робота з PostgreSQL
psycopg2-binary==2.9.9

# Безпека та паролі (для Групи 5)
Werkzeug==3.0.1

# Робота зі змінними оточення (для db_manager.py)
python-dotenv==1.0.0

# Додатково для авторизації (опціонально, але корисно)
Flask-Login==0.6.3